fix: emit_itoa_inline overflow on large code offsets
The backward jnz jump in emit_itoa_inline's div loop computed:
  let jmp_back = div_loop_pos as i8 - (code.len() + 2) as i8;

When div_loop_pos and code.len() are large (>127), the `as i8`
casts truncate BEFORE the subtraction, causing a signed overflow
panic in debug builds. The DIFFERENCE always fits in i8 (the div
loop is ~20 bytes), but the individual operands don't.

Fix: compute as i32 first, then cast the result.

Found via: examples/pricing.verbose::penalty_rate (nested if/else
in a rule with let bindings, which generates enough code to push
the itoa offset past 127).

25 of 29 examples now compile natively (up from ~20 at start
of session). Remaining failures: app (modules), clients (text
comparison), retirement (collection(bool) output).

All 159 tests pass.

+++ src/native.rs
@@ -4648,7 +4648,7 @@ fn emit_itoa_inline(code: &mut Vec<u8>) {
     // test rax, rax
     code.extend_from_slice(&[0x48, 0x85, 0xC0]);
     // jnz .div_loop
-    let jmp_back = div_loop_pos as i8 - (code.len() + 2) as i8;
+    let jmp_back = div_loop_pos as i32 - (code.len() + 2) as i32;
     code.extend_from_slice(&[0x75, jmp_back as u8]);
 
     // .write:
